requirements.txt
selenium
webdriver_manager
requests
python-dotenv
